Agentic AI Solution Design Patterns/search_correction/requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
